+++ CHANGELOG.md
@@ -36,7 +36,7 @@
 
 🥷 *Internal*
 * RayRuntime can now be configured to pass resources to underlying remote functions
-
+* Added version metadata to the workflow IR
 
 ## v0.44.0